revert changes to filters
