+++ containers/ecr-viewer/src/app/components/EcrFilters.tsx
@@ -1,8 +1,6 @@
 "use client";
 import React, { useEffect, useState } from "react";
 
-import { Label } from "@trussworks/react-uswds";
-
 import { useLibraryQueryParam } from "@/app/hooks/useQueryParam";
 import { formatDateTime } from "@/app/services/formatDateService";
 import {
@@ -12,12 +10,12 @@ import {
   DateRangeOptions,
   dateRangeLabels,
 } from "@/app/utils/date-utils";
-import { toKebabCase } from "@/app/utils/format-utils";
 
 import {
   Filter,
   RadioDateOption,
   RadioDateOptions,
+  CustomDateInput,
   SelectDeselectAllCheckbox,
   CheckboxOptions,
 } from "./BaseFilter";
@@ -167,45 +165,6 @@ const FilterReportableConditions = ({
  * - Updates the browser's query string when the filter is applied.
  */
 const FilterByDate = ({ initCustomDate, initDateRange }: FilterProps) => {
-  const CustomDateInput = ({
-    label,
-    onDateChange,
-    defaultValue,
-    isRequired,
-    minValue,
-  }: {
-    label: string;
-    onDateChange: (date: string) => void;
-    defaultValue: string;
-    isRequired: boolean;
-    minValue?: string;
-  }) => {
-    const today = new Date().toLocaleDateString("en-CA");
-    const id = toKebabCase(label);
-    return (
-      <div>
-        <Label htmlFor={id} className="margin-top-1">
-          {label}
-        </Label>
-        <input
-          id={id}
-          data-testid={id}
-          type="date"
-          className="usa-input width-card margin-top-0 border-base-dark"
-          defaultValue={defaultValue}
-          min={minValue}
-          max={today}
-          required={isRequired}
-          aria-label={label}
-          onChange={(e) => {
-            const date = e.target.value;
-            onDateChange(date);
-          }}
-        />
-      </div>
-    );
-  };
-
   const today = new Date().toLocaleDateString("en-CA");
   const { deleteQueryParam, updateQueryParam, pushQueryUpdate } =
     useLibraryQueryParam();
